// src/types/authorizer/v2IsRequest.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Authorizer
 * Authorizer service
 * OpenAPI spec version: v0.20.10
 */
import type { ApiIdentityContext } from './apiIdentityContext';
import type { ApiPolicyContext } from './apiPolicyContext';
import type { ApiPolicyInstance } from './apiPolicyInstance';
import type { V2IsRequestResourceContext } from './v2IsRequestResourceContext';

export interface V2IsRequest {
  identity_context?: ApiIdentityContext;
  policy_context?: ApiPolicyContext;
  policy_instance?: ApiPolicyInstance;
  resource_context?: V2IsRequestResourceContext;
}