// src/types/authorizer/v2QueryResponse.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Authorizer
 * Authorizer service
 * OpenAPI spec version: v0.20.10
 */
import type { V2QueryResponseMetrics } from './v2QueryResponseMetrics';
import type { V2QueryResponseResponse } from './v2QueryResponseResponse';
import type { V2QueryResponseTraceItem } from './v2QueryResponseTraceItem';

export interface V2QueryResponse {
  metrics?: V2QueryResponseMetrics;
  response?: V2QueryResponseResponse;
  trace?: V2QueryResponseTraceItem[];
  trace_summary?: string[];
}